pyproject.toml
[tool.poetry]
package-mode = false
name = "zoom-transcript-summariser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.3"
langchain-community = "^0.0.36"
torch = "^2.3.0"
transformers = "^4.40.1"
jupyterlab = "^4.1.8"
webvtt-py = "^0.4.6"
accelerate = "^0.30.1"
huggingface-hub = "^0.27.0"
ipywidgets = "^8.1.5"
load-dotenv = "^0.1.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
dynaconf = "^3.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
